requirements.txt
fastapi
uvicorn
python-dotenv
openai
google-cloud-vision
python-multipart